refactor(syntax): 绿树存储层 RawKind:int 化——语言无关地基
P1-E-1（token/节点/Lexer 分家的前提）：GreenNode 字段从 SyntaxKind 改为 int RawKind，
保留 Kind => (SyntaxKind)RawKind 便捷视图（过渡态）；GreenToken/GreenNodeWithChildren 构造内部转 int；
RedNode.Kind 经 _green.RawKind。存储层与语言枚举类型解耦，为拆两套 SyntaxKind（P1-E-2）铺路。
GreenNode 内部 100+ 处结构判断暂保持等价（行为零差异）。

新增 GreenNode_RawKind_MatchesKindStorage 契约测试锁定存储语义。
全量 41789 测试通过

+++ src/Cocoa.Cs/Cocoa.Core/Syntax/Green/GreenNodeWithChildren.cs
@@ -9,7 +9,7 @@ namespace Cocoa.CodeAnalysis.Syntax
         private readonly ImmutableArray<GreenNode?> _slots;
 
         public GreenNodeWithChildren(SyntaxKind kind, ImmutableArray<GreenNode?> slots)
-            : base(kind)
+            : base((int)kind)
         {
             _slots = slots.IsDefault ? ImmutableArray<GreenNode?>.Empty : slots;
         }
